:+1: Add types on `getwininfo` and `getwinpos` functions
